lint fixes and rename language files
