+++ src/react/components/shell/sidebar.tsx
@@ -36,12 +36,14 @@ export default function Sidebar({ project }) {
                     </ConditionalNavLink>
                 </li>
                 <li>
-                    <NavLink title={strings.connections.title} to={`/connections`}><i className="fas fa-plug"></i></NavLink>
+                    <NavLink title={strings.connections.title}
+                            to={`/connections`}><i className="fas fa-plug"></i></NavLink>
                 </li>
             </ul>
             <div className="app-sidebar-fill"></div>
             <ul>
-                <li><NavLink title={strings.appSettings.title} to={`/settings`}><i className="fas fa-cog"></i></NavLink></li>
+                <li><NavLink title={strings.appSettings.title}
+                            to={`/settings`}><i className="fas fa-cog"></i></NavLink></li>
             </ul>
         </div>
     );
